pyproject.toml
[project]
name = "olympic-boxing-video-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gdown>=5.2.0",
    "opencv-python-headless>=4.12.0.88",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pycocotools>=2.0.7",
    "torchinfo>=1.8.0",
    "tensorboard>=2.15.0",
    "nvidia-dali-cuda120>=1.35.0",
]
